Moved objects.csv function
