+++ Python/tdw/physics_audio/object_audio_static.py
@@ -1,3 +1,8 @@
+import io
+from csv import DictReader
+from pathlib import Path
+from pkg_resources import resource_filename
+from typing import Union, Dict
 from tdw.physics_audio.audio_material import AudioMaterial
 
 
@@ -51,3 +56,38 @@ class ObjectAudioStatic:
         The ID of the object.
         """
         self.object_id = object_id
+
+
+def get_static_audio_data(csv_file: Union[str, Path] = "") -> Dict[str, ObjectAudioStatic]:
+    """
+    Returns ObjectInfo values.
+    As of right now, only a few objects in the TDW model libraries are included. More will be added in time.
+
+    :param csv_file: The path to the .csv file containing the object info. By default, it will load `tdw/physics_audio/objects.csv`. If you want to make your own spreadsheet, use this file as a reference.
+
+    :return: A list of default ObjectInfo. Key = the name of the model. Value = object info.
+    """
+
+    objects: Dict[str, ObjectAudioStatic] = {}
+    # Load the objects.csv metadata file.
+    if isinstance(csv_file, str):
+        # Load the default file.
+        if csv_file == "":
+            csv_file = str(Path(resource_filename(__name__, f"objects.csv")).resolve())
+        else:
+            csv_file = str(Path(csv_file).resolve())
+    else:
+        csv_file = str(csv_file.resolve())
+
+    # Parse the .csv file.
+    with io.open(csv_file, newline='', encoding='utf-8-sig') as f:
+        reader = DictReader(f)
+        for row in reader:
+            o = ObjectAudioStatic(name=row["name"], amp=float(row["amp"]), mass=float(row["mass"]),
+                                  material=AudioMaterial[row["material"]], bounciness=float(row["bounciness"]),
+                                  resonance=float(row["resonance"]), size=int(row["size"]), object_id=0)
+            objects.update({o.name: o})
+    return objects
+
+
+DEFAULT_OBJECT_AUDIO_STATIC_DATA: Dict[str, ObjectAudioStatic] = get_static_audio_data()
